update tick type

+++ assembly/UniswapV3LiquidityStrategy.ts
@@ -1,11 +1,11 @@
 import { BaseStrategy } from '@steerprotocol/base-strategy/assembly';
 
 export class UniswapV3LiquidityStrategy extends BaseStrategy {
-  private LIQUIDITY_PRECISION: f64 = 2 ** 16 - 2; 
-  private startTick: f64;
-  public binWidth: f64 = 600;
+  private LIQUIDITY_PRECISION: i64 = 2 ** 16 - 2; 
+  private startTick: i32;
+  public binWidth: i32 = 600;
 
-  constructor(binWidth: f64) {
+  constructor(binWidth: i32) {
     super();
     this.binWidth = binWidth;
   }
